bio-match-platform/requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Web scraping
beautifulsoup4==4.12.2
lxml==4.9.3
scrapy==2.11.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# API framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3

# Database
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# Caching
diskcache==5.6.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
tqdm==4.66.1
click==8.1.7

# XML processing (for PubMed)
lxml==4.9.3

# Network analysis (optional)
networkx==3.2.1
matplotlib==3.8.2

# Data export
pyarrow==14.0.2

# Development tools
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
